requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Login==0.6.3
Flask-WTF==1.2.2
openpyxl==3.1.5
python-docx==1.1.2
python-dotenv==1.0.1
gunicorn==23.0.0
Werkzeug==3.1.3
email-validator==2.2.0
